fixes some essential banning stuff, few misc things

+++ Interstation-Two-WW2/code/modules/admin/verbs/debug.dm
@@ -569,6 +569,8 @@
 	set desc = "Removes ALL blood from the map"
 
 	for(var/B in blood)
-		qdel(B)
+		if (istype(B, /datum))
+			qdel(B)
+
 	log_admin("[key_name(usr)] has removed all blood.")
 	message_admins("<span class='notice'>[key_name_admin(usr)] has removed all blood.</span>", 1)
